requirements.txt
pyserial==3.5
svgpathtools==1.4.4